move  include of log utils up and rename impl files to sensible names
