+++ source/include/logging/impl/cs_LogPlainText.h
